# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Owner-scoped tasks, identity, wallet, sandbox payments and notifications.",
        "title": "FinTask API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/tasks/{id}": {
            "get": {
                "description": "Получение конкретной задачи по её ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получение задачи",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID получаемой задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача успешной найдена",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.GetTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить существующую в системе задачу по её ID",
                "tags": [
                    "tasks"
                ],
                "summary": "Удаление задачи",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID удаляемой задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешное удаление задачи"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Обновляет информацию об уже существующей в системе задаче\n### Логика обновления полей (Three-state logic):\n1. **Поле не передано**: `description` игнорируется, значение в БД не меняется\n2. **Явно передано значение**: `\"description\": \"Утром в 06:30 выйти на прогулку с Бобиком\"` — устанавливает новое описание для задачи\n3. **Явно передан null**: `\"description\": null`— очищает поле в БД (set to NULL)\nОграничения: `title` и `completed` не могут быть выставлены как null",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Обновить задачу",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID изменяемой задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PatchTask тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.PatchTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно изменённая задача",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.PatchUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v2/payments/webhook/yookassa": {
            "post": {
                "description": "Повторно проверяет платеж через API и выполняет одно зачисление",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Принять уведомление YooKassa",
                "parameters": [
                    {
                        "description": "Уведомление payment.succeeded",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_payments_transport_http.YooKassaWebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v2/tasks/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Возвращает задачу только текущего пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks-v2"
                ],
                "summary": "Получить свою задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http_v2.TaskResponse"
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http_v2.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http_v2.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CSRFToken": [],
                        "SessionCookie": []
                    }
                ],
                "description": "Удаляет задачу только текущего пользователя",
                "tags": [
                    "tasks-v2"
                ],
                "summary": "Удалить свою задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Задача удалена"
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http_v2.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http_v2.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CSRFToken": [],
                        "SessionCookie": []
                    }
                ],
                "description": "Применяет nullable patch только к задаче текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks-v2"
                ],
                "summary": "Изменить свою задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Изменяемые поля",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http_v2.PatchTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная задача",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http_v2.TaskResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный patch",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http_v2.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http_v2.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http_v2.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Конкурентное изменение",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http_v2.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_LisLisich_RESTAPI_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "message": {
                    "type": "string",
                    "example": "short human-readable message"
                }
            }
        },
...
        "internal_features_payments_transport_http.YooKassaWebhookRequest": {
            "type": "object",
            "required": [
                "event",
                "object",
                "type"
            ],
            "properties": {
                "event": {
                    "type": "string"
                },
                "object": {
                    "type": "object",
                    "required": [
                        "id"
                    ],
                    "properties": {
                        "id": {
                            "type": "string"
                        }
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "internal_features_tasks_transport_http.GetTaskResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 5
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completedAt": {
                    "type": "string",
                    "example": "null"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Сделать до четверга домашнее задание по математике"
                },
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "title": {
                    "type": "string",
                    "example": "Домашка"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "internal_features_tasks_transport_http.PatchTaskRequest": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string",
                    "example": "null"
                },
                "title": {
                    "type": "string",
                    "example": "Погулять с собакой"
                }
            }
        },
        "internal_features_tasks_transport_http.PatchUserResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 5
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completedAt": {
                    "type": "string",
                    "example": "null"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Сделать до четверга домашнее задание по математике"
                },
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "title": {
                    "type": "string",
                    "example": "Домашка"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "internal_features_tasks_transport_http_v2.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "message": {
                    "type": "string",
                    "example": "short human-readable message"
                }
            }
        },
        "internal_features_tasks_transport_http_v2.PatchTaskRequest": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "internal_features_tasks_transport_http_v2.TaskResponse": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access JWT в формате: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CSRFToken": {
            "description": "CSRF-токен для изменяющих запросов с cookie-сессией",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        },
        "SessionCookie": {
            "description": "Защищенная cookie-сессия браузера",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}